call #set_lower_submittable_to_be_destroyed periodically from worker process
The following case happens when run#to_be_destroyed is set true while job_observer is creating auto-run analyses
- run (to_be_destroyed: true)
  - analysis (to_be_destroyed: false)

If this happens, run and analysis is never destroyed because the run
is waiting the analysis to be destroyed.
In order to avoid this situation, run#set_lower_submittable_to_be_destroyed
should be called periodically by the worker.
This is not limited to the case of runs, but for all other documents
such as simulator, parameter_set, and analyzer.
So we decided to call it in job_observer, job_submitter, and document_destroyer