fix a possible cause of the ajax error #514
See https://github.com/crest-cassia/oacis/issues/514
A possible cause of this problem is that the list of ParameterSets are
not fixed, but the query is sustained.
If the status of a record changed during DataTable class is being
processed, the contents of the parameter_sets_list can change, which
causes an inconsistency. To avoid this problem, we fix the list of
ParameterSets by calling `to_a` to the query.